Fix bug in partitioning so that it checks if the file already exists
Also add a print so that the user visibly can see which part of the
code is running. (Since the zipping of the downloaded dataset can be
quite long)

+++ dataset/data_loader.py
@@ -99,6 +99,7 @@ def partition_dataset(subsets, classifications, train, val, test, images):
               'Val =', val, '\n'
               'Test =', test)
     else:
+        print('Start partitioning data...')
         subset_regex = '(' + ('|'.join(subsets)) + ')'
         for img in images:
             partition = random.randrange(100)
@@ -109,7 +110,8 @@ def partition_dataset(subsets, classifications, train, val, test, images):
             else:
                 new_set = 'test'
             new_path = re.sub('.+' + os.path.sep + subset_regex, new_set, img)
-            shutil.copy(img, new_path)
+            if not os.path.exists(new_path):
+                shutil.copy(img, new_path)
 
 
 if __name__ == '__main__':
